agents/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
fastapi>=0.104.0
pydantic>=2.0.0
textblob>=0.17.1
scikit-learn>=1.3.0
numpy>=1.24.0